% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_sf.R
\name{convert_to_sf}
\alias{convert_to_sf}
\title{Turn every activity into an sf object. Nest the data frame by 'id' and
'act_date_chr' to do this.}
\usage{
convert_to_sf(df_act_meas)
}
\arguments{
\item{df_act_meas}{}
}
\value{
sf object
}
\description{
Turn every activity into an sf object. Nest the data frame by 'id' and
'act_date_chr' to do this.
}
