% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_poi.R
\name{extract_poi}
\alias{extract_poi}
\title{Extract points of interest.}
\usage{
extract_poi(sf_act_meas)
}
\arguments{
\item{sf_act_meas}{}
}
\value{
Data frame with points of interest
}
\description{
Extract points of interest.
}
