% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpen_map.R
\name{get_alpen_map}
\alias{get_alpen_map}
\title{Get the ground map for the visualisation of the spatial data.}
\usage{
get_alpen_map(sf_act_meas, tol_bbox = 0.015, map_zoom = 10)
}
\arguments{
\item{map_zoom}{}
}
\value{

}
\description{
Get the ground map for the visualisation of the spatial data.
}
