% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_meas.R
\name{pre_process_meas}
\alias{pre_process_meas}
\title{Function to preprocess measurements of the included activities.}
\usage{
pre_process_meas(df_act_meas)
}
\arguments{
\item{df_act_meas}{}
}
\value{
Preprocessed activities
}
\description{
Function to preprocess measurements of the included activities.
}
\examples{
pre_process_meas(df_act_meas)
}
