% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_altitude_ridge.R
\name{vis_altitude_ridge}
\alias{vis_altitude_ridge}
\title{Visualise the altitude by using a ridge plot.}
\usage{
vis_altitude_ridge(df_act_meas)
}
\arguments{
\item{df_act_meas}{}
}
\value{
ggplot of altitude data
}
\description{
Visualise the altitude by using a ridge plot.
}
